Fixed Starfield Background
It's gonna say tons of meta files were affected because I had to try to fix the issue with all the scripts suddenly malfunctioning. I fixed the starfield background, couldn't get enemy projectiles to reappear, but added up to page 607 from Chapter 31 in the book. Spread firing does not work, and the Enemy_02 and Enemy_03 movements as given in Chapter 31 do not work and as such are set to the default enemy movement.